requirements.txt
SpeechRecognition
google-generativeai
edge-tts
playsound
pyaudio
